# Optimize subscription path for `FixedStreamMessage`