--- a/core/src/main/java/com/linecorp/armeria/common/FixedHttpRequest.java
+++ b/core/src/main/java/com/linecorp/armeria/common/FixedHttpRequest.java
@@ -19,6 +19,7 @@ package com.linecorp.armeria.common;
 import com.linecorp.armeria.common.stream.EmptyFixedStreamMessage;
 import com.linecorp.armeria.common.stream.OneElementFixedStreamMessage;
 import com.linecorp.armeria.common.stream.RegularFixedStreamMessage;
+import com.linecorp.armeria.common.stream.ThreeElementFixedStreamMessage;
 import com.linecorp.armeria.common.stream.TwoElementFixedStreamMessage;
 
 /**
@@ -68,6 +69,23 @@ final class FixedHttpRequest {
         TwoElementFixedHttpRequest(
                 RequestHeaders headers, HttpObject obj1, HttpObject obj2) {
             super(obj1, obj2);
+            this.headers = headers;
+        }
+
+        @Override
+        public RequestHeaders headers() {
+            return headers;
+        }
+    }
+
+    static final class ThreeElementFixedHttpRequest
+            extends ThreeElementFixedStreamMessage<HttpObject> implements HttpRequest {
+
+        private final RequestHeaders headers;
+
+        ThreeElementFixedHttpRequest(
+                RequestHeaders headers, HttpObject obj1, HttpObject obj2, HttpObject obj3) {
+            super(obj1, obj2, obj3);
             this.headers = headers;
         }