--- a/core/src/main/java/com/linecorp/armeria/common/HeaderOverridingHttpRequest.java
+++ b/core/src/main/java/com/linecorp/armeria/common/HeaderOverridingHttpRequest.java
@@ -84,11 +84,6 @@ final class HeaderOverridingHttpRequest implements HttpRequest {
     }
 
     @Override
-    public void subscribe(Subscriber<? super HttpObject> subscriber, EventExecutor executor) {
-        delegate.subscribe(subscriber, executor);
-    }
-
-    @Override
     public void subscribe(Subscriber<? super HttpObject> subscriber, EventExecutor executor,
                           SubscriptionOption... options) {
         delegate.subscribe(subscriber, executor, options);