--- a/core/src/main/java/com/linecorp/armeria/common/multipart/MultipartEncoder.java
+++ b/core/src/main/java/com/linecorp/armeria/common/multipart/MultipartEncoder.java
@@ -97,11 +97,6 @@ final class MultipartEncoder implements StreamMessage<HttpData> {
     }
 
     @Override
-    public void subscribe(Subscriber<? super HttpData> subscriber, EventExecutor executor) {
-        subscribe(subscriber, executor, EMPTY_OPTIONS);
-    }
-
-    @Override
     public void subscribe(Subscriber<? super HttpData> subscriber, EventExecutor executor,
                           SubscriptionOption... options) {
         requireNonNull(subscriber, "subscriber");