--- a/core/src/main/java/com/linecorp/armeria/common/stream/ConcatPublisherStreamMessage.java
+++ b/core/src/main/java/com/linecorp/armeria/common/stream/ConcatPublisherStreamMessage.java
@@ -15,7 +15,6 @@
  */
 package com.linecorp.armeria.common.stream;
 
-import static com.linecorp.armeria.common.stream.StreamMessageUtil.EMPTY_OPTIONS;
 import static com.linecorp.armeria.common.stream.StreamMessageUtil.containsNotifyCancellation;
 import static java.util.Objects.requireNonNull;
 
@@ -81,11 +80,6 @@ final class ConcatPublisherStreamMessage<T> implements StreamMessage<T> {
     @Override
     public CompletableFuture<Void> whenComplete() {
         return sources.whenComplete();
-    }
-
-    @Override
-    public void subscribe(Subscriber<? super T> subscriber, EventExecutor executor) {
-        subscribe(subscriber, executor, EMPTY_OPTIONS);
     }
 
     @Override