--- a/core/src/main/java/com/linecorp/armeria/common/stream/EmptyFixedStreamMessage.java
+++ b/core/src/main/java/com/linecorp/armeria/common/stream/EmptyFixedStreamMessage.java
@@ -16,8 +16,6 @@
 
 package com.linecorp.armeria.common.stream;
 
-import java.util.NoSuchElementException;
-
 import javax.annotation.Nullable;
 
 import com.linecorp.armeria.common.annotation.UnstableApi;
@@ -28,20 +26,14 @@ import com.linecorp.armeria.common.annotation.UnstableApi;
 @UnstableApi
 public class EmptyFixedStreamMessage<T> extends FixedStreamMessage<T> {
 
-    // No objects, so just notify of close as soon as there is demand.
-    @Override
-    final void doRequest(SubscriptionImpl subscription, long unused) {
-        if (requested() != 0) {
-            // Already have demand so don't need to do anything.
-            return;
-        }
-        setRequested(1);
-        notifySubscriberOfCloseEvent(subscription, SUCCESSFUL_CLOSE);
-    }
-
     @Override
     public final boolean isEmpty() {
         return true;
+    }
+
+    @Override
+    public long demand() {
+        return 0;
     }
 
     @Override
@@ -50,12 +42,8 @@ public class EmptyFixedStreamMessage<T> extends FixedStreamMessage<T> {
     }
 
     @Override
-    public boolean hasNext() {
-        return false;
-    }
+    public void request(long n) {}
 
     @Override
-    public T next() {
-        throw new NoSuchElementException();
-    }
+    public void cancel() {}
 }