--- a/core/src/main/java/com/linecorp/armeria/common/stream/FixedStreamMessage.java
+++ b/core/src/main/java/com/linecorp/armeria/common/stream/FixedStreamMessage.java
@@ -63,7 +63,6 @@ abstract class FixedStreamMessage<T> implements StreamMessage<T>, Subscription {
     @Nullable
     private volatile EventExecutor executor;
 
-    // Updated only via abortCauseUpdater
     @Nullable
     private volatile Throwable abortCause;
     // Updated only via subscribedUpdater