Motivation:

This is proof of concept to optimize subscription of `FixedStreamMessage`.
Most RESTful requests do not use a streaming request and response.
`FixedStreamMessage` is a crucial part of the performance for a unary request and response.

`FixedStreamMessage` holds all data when it is created.
We can minimize call stacks depending on their data size in order to improve overall performance.

Modification:

  - Remove `AbstractStreamMessage` from `FixedStreamMessage` for de-generalization and and optmized code path.
  - Directly implement `Subscription` by the subclasses of `FixedStreamMessage` for less object allocations.
  - Optimize `cancel()` and `abort()` depending of the size of elements.
  - Fix an possible infinity recursion in `DefaultStreamMessage` when using a direct executor.
  - Change `StreamMessage.subscribe(Subscriber, EventExecutor)` into a `default` implementation.

Benchmark result:

The optimized `FixedStreamMessage` shows 4 to 4.7% enhanced performance with regard to a simple GET request.
- Working branch
  ```
  wrk -t12 -c400 -d30s http://127.0.0.1:8080/foo
  Running 30s test @ http://127.0.0.1:8080/foo
    12 threads and 400 connections
    Thread Stats   Avg      Stdev     Max   +/- Stdev
      Latency    17.00ms   29.67ms 298.04ms   86.43%
      Req/Sec    10.50k     1.34k   32.03k    78.52%
    3764342 requests in 30.10s, 588.75MB read
    Socket errors: connect 0, read 201, write 0, timeout 0
  Requests/sec: 125058.64
  Transfer/sec:     19.56MB
  ```
- Master branch
  ```
  wrk -t12 -c400 -d30s http://127.0.0.1:8080/foo
  Running 30s test @ http://127.0.0.1:8080/foo
    12 threads and 400 connections
    Thread Stats   Avg      Stdev     Max   +/- Stdev
      Latency    17.63ms   30.47ms 257.90ms   86.10%
      Req/Sec    10.03k     1.72k   63.78k    83.45%
    3595215 requests in 30.10s, 562.30MB read
    Socket errors: connect 0, read 91, write 0, timeout 0
  Requests/sec: 119439.48
  Transfer/sec:     18.68MB
  ```