@minwoox 2021-05-26
> Benchmark result:
The implementation has been changed. TBU

Are you going to work on it?
Can I review now anyway? 😄