@anuraaga 2021-05-26
Thanks for all the optimizations @ikhoon - just wondering if you've had a chance to check gRPC performance lately as in #799? We have benchmarks for direct comparison with upstream and have seen largish gaps so IMO it's better marketing for Armeria to close that as much as possible to make sure it's a drop-in replacement. I have a feeling focusing on that gap would improve the overall framework performance too generally. As gRPC uses streaming requests, I guess targeting fixed streams doesn't help with it unfortunately.