@ikhoon 2021-05-26
Actually, in the previous optimization PRs, I used the CPU frame graphs ran with https://github.com/LesnyRumcajs/grpc_bench to optimize the gRPC performance. 😄

> As gRPC uses streaming requests, I guess targeting fixed streams doesn't help with it unfortunately.

Yeah, `DefaultStreamMessage` is the mission critical part of Armeria's general performance on streaming. 
I'd like to add a variant of `DefaultStreamMessage` that is a non-thread safe and could work on a single event loop for the internal implementations in the safe zone.