@ikhoon 2021-05-26
> As gRPC uses streaming requests, 

Filed an issue for this. #3587