@ikhoon 2021-05-26
The final benchmark result is updated to the PR description.